SERVER-93655 Extend query result comparator in shell_utils for QueryTester normalization (#26211)
[SERVER-93655](https://jira.mongodb.org/browse/SERVER-93655) added an
unordered results comparison in c++ to the shell_utils library. While
the original PR introduced a field-order agnostic sort, it didn't handle
sorting arrays (and nested arrays) with a `BSONObj`. This ticket extends
the helper to allow an option for sorting arrays, adding an `arrayCmp`
comparator function that considers arrays greater than non-arrays.

Additionally, this PR adds support for normalizing numeric types such
that equivalent numbers in different types and representations will be
considered equivalent. For instance, `NumberInt(1)` and
`NumberDecimal('1.00')` will be normalized to the same type and
sanitized of trailing zeroes.

This PR is a dependency for
[DEVPROD-9636](https://jira.mongodb.org/browse/DEVPROD-9636) - Normalize
`QueryTester` output. The functions required by that PR are put into
`shell_utils.h` so they can be accessed by `QueryTester`.