SERVER-90739 Fix selective restores for time-series collections (#26182)
This fixes sharded selective restores for time-series collections.

When creating a time-series collection with namespace `abc`, a view
(`abc`) and collection (`system.buckets.abc`) are created. The view
points to the collection. The `$backupCursor`, which reports the files
to be copied for a backup includes an `ns` field. The `ns` field is
reporting time-series collections using the view namespace and not the
bucket namespace, which is hidden from the user. However, the config
server collections are keyed by the buckets namespace. Part of the
selective restore procedure for the config server is to create a
collection in the local database called `system.collections_to_restore`.
This collection is populated with the namespaces and uuids of
collections to restore. I.E:
```
{
    ns: <string>,
    uuid: <UUID>
}
```

Cloud inserts the view namespace into this collection, as that's what we
reported. We can make a minimal change to
`configsvr_run_restore_command.cpp` to fix this.