SERVER-93649 shared_ptr for all TaskExecutor instances (#26357)
`TaskExecutor` needs to be managed by `shared_ptr` at all times, to be
ready for `shared_from_this` calls.

All classes derived from `TaskExecutor` hereby hide their constructors
with a passkey idiom, and must be created with a `create` static member
functions.

An exception to the rule is made for `TaskExecutorProxy` subclasses, as
there are several trivial subclasses of it, it's not production code,
and their uses are near their definitions, and are easily auditable.