SERVER-92675 Add a fuzzer for the BSONColumn builder (#26375)
The fuzzer seems to run for awhile without finding bugs now, so bringing
back the fuzzer PR.

The comment above `createFuzzedElement` summarizes the thought process
for how to generate BSONElements to construct inputs to the
BSONColumnBuilder, would like specific attention here if anyone has any
thoughts/concerns about how I'm doing this.

There is a unit test to go with this fuzzer that takes fuzzed strings
and replicates the results for purposes of reproducing them in gdb. I'm
not including it in this PR as I want to take some time to think over
what needs to be moved to bson_test_util to reduce code duplication.