SERVER-90983 Separate service arch owned files to new subteams (#26262)
This change removes the old service architecture team as a codeowner and
replaces it with the new server programmability, workload scheduling,
and networking and observability teams.

For the most part, this just involved changing
server-service-architecture to one of the new teams. A few exceptions
are:

1. src/mongo/executor - this directory is now split between NO and SP,
while it was previously all SA.
2. src/mongo/db/concurrency/flow_control_ticketholder* - previously not
owned, I added ownership for WS
3. src/mongo/db/dbdirectclient* - previously not owned, I added
ownership for SP given that src/mongo/scripting/dbdirectclient_factory.*
was owned by SA (and is now owned by SP).