SERVER-94223 WiredTigerEventHandler correctly waits for active readers to finish (#26630)
The precondition in this wait is backward. We need to wait until
`_activeSections == 0`.

In practice, this could be a problem that would allow storage engine
shutdown to incorrectly proceed when there is an active reader, but the
GlobalLock in the serverStatus section, which is the only user of this
mechanism, protects us from this case.