SERVER-87364 Replace LogoutCommandState Decorator atomic bool with std::call_once (#23453)
LogoutCommandState is a ServiceContext decoration that stores an atomic
bool used to track whether logout has been called. logout uses this to
emit deprecation warnings.

Decorations are powerful but complicated. This PR is to replace all of
this with a static std::once_flag and a call to std::call_once.

Co-authored-by: Gabriel Marks <gabriel.marks@mongodb.com>