SERVER-89535 Validate OCSP Stapling nextUpdate time to fit within a 64-bit size limit (#23895)
[Server Ticket](https://jira.mongodb.org/browse/SERVER-89535)
[Help Ticket](https://jira.mongodb.org/browse/HELP-58238)


## Overview
As stated in the above server ticket, if OCSP stapling is enabled, the
server starts a periodic job to fetch OCSP status from the responder.
Unless `ocspStaplingRefreshPeriodSecs` is configured to have a shorter
duration, the OCSP fetcher will use a duration calculated from the OCSP
response's `nextUpdate` field (if it has one).

This duration is converted to nanoseconds in the PeriodicJobImpl for the
next execution of the fetch job, which resulted in `DurationOverflow:
Overflow casting from a lower-precision duration to a higher-precision
duration`. The overflow, in turn results in server crash by way of an
unhandled uassert.

## Fix
When constructing a OCSPFetchResponse, the upper limit will be 60 days.
This will ensure that the nextUpdateTime will always be less than or
equal to 60 days, which would not overflow in nanoseconds.

---------

Co-authored-by: Gabriel Marks <gabriel.marks@mongodb.com>