SERVER-86376 The '.watch()' methods in mongo (legacy) shell modify the first argument (#24315)
This PR solves
[SERVER-86376](https://jira.mongodb.org/browse/SERVER-86376). These
changes replace the .aggregate() 'pipeline' argument (type is an array)
with a new array that has the changeStreamStage as its first element and
the stages of the initial pipeline as its subsequent elements (by using
the Spread operator on the initial pipeline). This way the pipeline
function parameter is not modified, but a copy of the updated pipeline
is used for the return function call.

---------

Co-authored-by: Ivan Fefer <ivan.fefer@mongodb.com>