SERVER-90335 Don't use a NLJ stage to build local keys in $lookup (#22389)
# Issue

SERVER-90335

# Description

The SBE microcode for extracting the set of keys from a stream that will
be matched against the keys extracted from the foreign collection in a
$lookup instruction uses a NLJ stage and a complex sequence of stages
doing getField/unwind/getField/unwind followed by a group stage that
stashes the leaf nodes in the final set, plus a "if empty, insert null".
Apart from being slow because of the presence of multiple stages to
unpack and re-pack the results, it's going to make the vectorization of
$lookup harder (we would need to have a vectorized NLJ stage). This
ticket started as an attempt to see if we could migrate this structure
into a single project stage to apply on top of the main stream.
E.g. from 

```
        [2] nlj inner [s1] [s1] 
            left 
                [1] scan s1 s2 none none none none none none lowPriority [] @\"afc8b24e-f451-42e6-a722-3cecca2032a0\" true false 
            right 
                [2] project [s10 = 
                    if isArrayEmpty(s8) 
                    then [null] 
                    else s8 
               ] 
                [2] group [] [s8 = addToSet(s6)] spillSlots[s9] mergingExprs[aggSetUnion(s9)] 
                [2] unwind s6 s7 s5 true 
                [2] project [s5 = getField(s1, \"shardKey\")] 
                [2] limit 1ll 
                [2] coscan 
```
to
```
        [2] project [s5 = arrayToSet(
            let [
                l101.0 = getField(s1, \"shardKey\") 
            ] 
            in 
                if (isArray(l101.0) ?: false) 
                then 
                    if isArrayEmpty(l101.0) 
                    then [null] 
                    else move(l101.0) 
                
                else newArray((move(l101.0) ?: null)) 
       )] 
        [1] scan s1 s2 none none none none none none lowPriority [] @\"08a693ed-8215-4b68-ab31-1d2da2b3739a\" true false 
```

---------

Co-authored-by: Ivan Fefer <ivan.fefer@mongodb.com>