SERVER-90965 Extend query settings to contain optional comment field (#24738)
This PR implements
[SERVER-90965](https://jira.mongodb.org/browse/SERVER-90965). These
changes extend the `QuerySettings` with an optional `Comment`, an
improvement that will allow users to insert / update the comment, which
is of `any` type, of any query settings instance.
This extension is implemented by adding a 2 new files:
- `query_settings_comment.h`
- `query_settings_comment.cpp`

and by adding the `Comment` field to `query_settings.idl` and
`idl_check_compatibility.idl`.
These changes were tested via `.cpp` unit tests and `.js` end-to-end
tests.

---------

Co-authored-by: Ivan Fefer <ivan.fefer@mongodb.com>