SERVER-84089 Check for collation when retrying bulk write upsert (#26578)
In a bulk write if we perform single upsert operation with the query
that exactly matches a unique index, we are allowed to retry it. See
[write_ops_exec::shouldRetryDuplicateKeyException](https://github.com/mongodb/mongo/blob/2365d60a326c0a0ebfc4be5e602a9cb5acba0a30/src/mongo/db/ops/write_ops_exec.cpp#L2230).

However, if the collation of the unique index and a query doesn't match,
this is may be incorrect and lead to an infinite retry loop.