SERVER-94375 Make count10.js more robust against QueryPlanKilled (#26701)
`count10.js` assumes that the count operation will fail when interrupted
by the parallel shell. However, in suites with balancer, the balancer
can kill the operation with `QueryPlanKilled`. When this happens, the
parallel shell will never find the count operation, and the test will
hang and eventually timeout. This PR fixes this situation by retrying as
well as doing so more frequently to beat the balancer.