SERVER-93059 Remove unnecessary projection for distinct (#26658)
When integrating multiplanning with DISTINCT_SCAN, we added logic to
include a projection all the time to ensure we preserved correctness.
However, we should only add a projection when needed to return output
from a covered distinct scan- otherwise, if we have a FETCH stage, we
don't need a projection for distinct queries.