SERVER-89033 Create plan cache targeted fuzzer (#22590)
Creates a new fuzzer-like integration test for plan caching. This
operates by generating random documents and aggregation queries. It runs
each agg query twice as originally created, so that if caching is
enabled it will become cached and active (if cache-eligible), then
tweaks a parameter value in it and runs it again, comparing the first to
the final result.

This does not test whether caching occurs when expected; it only tests
whether caching, if and when it happens, produces the same result as
non-cached execution. Thus it does not need to be updated if the caching
criteria change and is agnostic to whatever cache algorithm is active or
whether there is any caching being done at all. It is also capable of
finding bugs that do not involve the plan cache, such as hitting
tasserts, and since it was written independently of other fuzzers it is
more likely to expand coverage into areas existing fuzzers have missed.

It supports reproducing a failure either by rerunning just the failing
query or the entire test.

Currently it supports aggregation stages $group, $lookup, $match,
$project, $sort, and $unwind. These are all those that were planned for
initial delivery. More can be added incrementally in the future if
desired.