SERVER-93738 Excludes tests tagged with newer versions of mongot (#26776)
Adds the ability to tag tests with a specific version of mongot that is
required in order to run the test.

Due to how resmoke works, we don't directly compare the tag of a test to
the mongot version, and we instead exclude tags based on the version of
mongot running. Given a version of mongot, we exclude any tags for the
next API version, the next 5 releases, and the next 10 patch releases.

As you would only want to exclude tests if the changes are in latest
mongot, but not in the release mongot, the changes should be in an
imminent future release. These values could be modified in the future if
necessary.