SERVER-94042 move amazon linux require variant to static builds (#26680)
This saves time running tests and the new distro saves time compiling. I
made unittests stay as dynamic because they do not need the improvement.
The main downside after this change is that
compile_all_but_not_unittests takes a long time to complete. It is not
in any dependency chains so this does not matter too much but if we have
a solution that can fix this we should do it.