SERVER-83235 Implement anti-materialization in the SBE stage builder (#25443)
This PR changes the SBE stage builder to implement the new
anti-materialization approach.

In addition, this PR also makes a few other changes:

1) I improved ProjectionEffects (adding new effects 'kSet' and 'kAdd')
and renamed it to FieldEffects.

2) I added a generateSingleFieldProjection() function so we could apply
the relevant parts of a projection to an individual field, without
having to compute the result of applying the entire projection to the
full result document.

3) I removed the "input fields" feature from makeBsonObj(), and I also
removed the MakeObjInputPlan, InputFieldsOnlyCursor,
BsonObjWithInputFieldsCursor, ObjWithInputFieldsCursor, and
MakeObjCursorInputFields classes. These features made sense for the
older anti-materialization design, which was going to involve having
dynamic state at run time to keep track of field order. However, after
some discussion with members of the QE team we decided we didn't want to
go with an approach that required dynamic state at run time to keep
track of field order. Given this, all of these features were no longer
necessary.

4) I took some of the anti-materialization decision logic from
makeBuildProjectionPlan() and I split it out into a separate
getResultPlan()/makeResultUsingPlan() mechanism. This made the code much
simpler and easier to reason about.

For testing, I'm relying on the "multiple_projects.js" and
"complex_pipelines.js" tests we have in master, which should give pretty
good coverage both for correctness, and for ensuring that field order
isn't affected by anti-materialization.