SERVER-93576: Redacted queryVector Embeddings Field and Added Test (#26882)
Redacted the queryVector embeddings field for explain objects if it
exists. Added a jstest called
jstests/with_mongot/e2e/vector_search_remove_embeddings.js to check that
the field is properly redacted and confirmed it passes single node and
sharded cluster test suites.

For time being, the queryVector embeddings field value is replaced with
"redacted."

Observed that the jstest sometimes fails burn_in task for single node
test suite (when the test is invoked repeatedly) but the failing
behavior does not appear to be related to the made code changes. A
ticket for this issue has been created (SERVER-94633)