SERVER-94564 Increase evg timeouts for blockprocessing and change stream fuzzers in additional build variants (#26825)
The block processing fuzzer runs on additional build variants in
`mongodb-mongo-master-nightly`, so we should increase the timeouts there
to resolve [BF-34912](https://jira.mongodb.org/browse/BF-34912). The
change stream fuzzers also need timeout increases in the
`enterprise-amazon2023-arm64-fuzzers` variant.

This is a follow-up to
[SERVER-94484](https://jira.mongodb.org/browse/SERVER-94484) following
recent changes to the fuzzer grammar as part of
[SPM-3714](https://jira.mongodb.org/browse/SPM-3714).