SERVER-94538: Increase txn timeout in multi_stmt_txn_jscore_passthrough_with_migration (#26899)
Attempt to address BF-34664.

Limit transaction retries in tests not by a maximum number of retries
*and* a timeout, but only limit it by a timeout.
That will allow an arbitrary number of retries, but still keep the test
time bounded by the timeout. It reduces potential test failures due to
the maximum number of retries being exhausted although the timeout isn't
reached.

The PR also increases the total timeout for the affected test from 5
minutes to 10 minutes.
The `RetryTracker` is also slightly adjusted in this PR, so it does not
have the ability anymore to limit the number of retries.