SERVER-93834 Path decompressor invariant permits non-empty, but scalarless obj literals (#26879)
The decompress paths fuzzer passed a string that contained an object
literal that had some nesting, but no scalars, to the path decompressor.
This relaxes the invariant to permit this, as it should not be
problematic. Traversal looking for scalars is gated behind the initial
check for empty object to avoid doing the work where possible (and this
should eventually be removed when we can evaluate paths in literals.