SERVER-94674 Make spawnhost authenticate with ec2 metadata service (#26932)
Some distros have different versions of the metadata service running on
the ec2 instance. This accounts for both versions and tries to
authenticate. If acquiring a token fails we just try the api without
authenticating afterwards.

I tested this on one of the distros that used the old version of the
metadata service and the distro that caused the issue. This fixed it.