SERVER-91458 Update mirrored reads metrics to consider "erroredOnRead" case (#26671)
To accurately increment the `sent` metric, maybe we could:
1) Add something to "sending" (AsyncDBClient/Session). There are some
connection health metrics there, but none that could be used for our
purpose. Also probably not desirable to implement something here just
for a mirrored reads metric.
2) Inspect the `Status` we get after schedule. I think this is wrong
given that if a network timeout error like a connection acquisition
timeout occurs after schedule, we would get ok status but the request
never sends.
3) We could also inspect the `Status` we get from the response future,
but that would be effectively the same a `resolved`.

For the reasons as above, it would probably be difficult to accurately
increment the `sent` metric. I think it's more accurate to instead just
rename the metric to `scheduled`, and include a `failedToSend` metric
which considers sending errors that `sent` failed to consider. `pending`
is changed similarly.

### Changes
* `sent` -> `scheduled`, metric incremented in a different location,
test changed accordingly
* `failedToSend` metric added, add test for failed to send case
* `pending` incremented and decremented in a different location, test
changed accordingly