SERVER-93983 add consolidated target support (#26887)
This change setups up a basic foundation for a framework to optionally
build certain binaries in a consolidated fashion via a command line arg.
This useful to help preserve diskspace in evergreen where we want to
link many unittests binaries statically. This allows those many to be
consolidated into a lesser number.

This change only consolidates a few binaries, just enough to clear the
BF. Also not any binary will be easily consolidated with any other
binary. The binaries should share a common environment, have similar
deps (non conflicting) and in general make sense for consolidation.