SERVER-94740 CheckMetadataConsistency can trigger false positives due to looking at sharding metadata while critical section is active (#27099)
CheckMetadataConsistency uses the sharding metadata installed on the
shards to determine if a shard currently owns chunks of a collection.
And based on that, it determines if the collection should be present on
the local catalog and flags an inconsistency if it isn't.

CheckMetadataConsistency reads the sharding metadata by using
CollectionShardingRuntime::getCurrentMetadataIfKnown, which ignores the
critical section state – it simply returns the currently installed
metadata, even if the critical section is active. If it is active, then
the currently installed metadata might not be the valid one, since it is
currently being changed. This can cause CheckMetadataConsistency to
report false positives.

Considering CheckMetadataConsistency is based on a best effort approach,
in case of critical section active we will skip the check.