SERVER-91235 replace usages of bson manipulating functions in cluster… (#26845)
SERVER-91235
This commit replaces most instances of appendDbVersionIfPresent and
applyReadWriteConcern with similar functions that operate on an idl
struct. In both cases, there were some instances that could not be
removed because the command derived from BasicCommand which does not
have an idl struct associated with it. Therefore, the
appendDbVersionIfPresent and applyReadWriteConcern functions must still
exist.