SERVER-94866 Replace parallel shell with thread (#27173)
On slow variants spawning shells can take a long time to load and
execute. In this test the slow execution of the shell contributes to
transaction lifetime exceeded errors causing unintended test failures.
We replace parallel shells with lightweight Thread objects in this PR
that are much faster to execute than shell processes.