SERVER-94502 Make RouterRole loops that reenter directly into the ShardRole not retry upon stale shard (#26876)
Problem: when a ShardRole operation transitions into the RouterRole and
then reenters back into the ShardRole, the shard might have unknown or
stale filtering metadata and this causes a StaleConfig exception. If the
innermost ShardRole is not wrapped with a shard role retry loop, then
the shard will not refresh its filtering metadata, but the RouterRole
will retry again only to hit the same error over and over.

Commands sent over the network are wrapped with a shard role retry loop,
because the ServiceEntryPoint implements it. However, there are a few
cases where as an optimization, we don't go over the network, and
instead reenter into the shardRole directly withing the same opCtx. In
this case, there's no shardRole retry loop.

This PR adds a configuration to the RouterRole::CollectionRouter so that
these anomalous usages can be flagged and make the CollectionRouter
bubble up the stale shard exception so that it is handled by a shardRoop
at the root of the operation. This servers as a low-risk stop-gap
solution that can be backported to v8.0, while a more definitive fix
will be done by SERVER-77402 by wrapping these operations with a
shardRole retry loop.