SERVER-73641 Only generate timeseries _id comparisons on mongod (#26408)
SERVER-73641 was a wrong answer bug where we would omit extended range
events for sharded timeseries collections. The reason for this was:
- The view for the timeseries collection can only be resolved on the DB
primary
- `mongos` plans the query, and and if `usesExtendedRange` for the
collection was false on the DB primary, then it will generate a
predicate on `_id` to help with index utilization.
- If the target shard has extended range events, they may be omitted
since a scan of `_id` may not find them.

The fix I implemented here was to always assume that there are extended
range events when planning on `mongos`. I also moved generation of `_id`
predicates into a separate method, so they can only be performed on
`mongod` for the target shards where it can be known that it's safe to
do so.

Some tests needed to be updated since `_id` predicates are no longer
generated in some cases:
- When creating partial indexes predicated on the time column
- When doing updates
In these cases we don't have info on whether or not the shard for the
collection contains extended range values, so we have to be
conservative. If this causes performance to regress for some queries,
the workaround for this is to create an index on the time column. It's
assumed that extended range values are quite infrequent, so hopefully
this is never an issue.