SERVER-93675: Make semantic_analysis consider renamed-over fields modified (#27154)
`$addFields::getModifiedPaths` for a query like
`{"$addFields":{"b":"$a"}}` reports
`GetModPathsReturn::Type::kFiniteSet` with rename `"a"->"b"`, but does
not include `"b"` as a modified path.

A range of tests confirm this was the intended behaviour, rather than an
oversight.

However, as a result semantic analysis erroneously assumed `"b"` is
_preserved_ when walking forwards through a pipeline; it is not, it has
been overwritten.

`GetModPathsReturn.paths` for `kFiniteSet` is effectively defined as
"this exact value does not exist on the other side of this operation, at
all" - that is when traversing in _either_ direction, this stage should
be considered to alter the reported fields.

These semantics are depended on widely - including `"b"` as a modified
path breaks many assumptions and optimisations, with a large fallout.
Stages reporting a field in `paths` _and_ as the destination of a rename
would (currently) prevent the rename being used for optimisations which
reorder stages.

`semantic_analysis` itself even relies on this elsewhere; it checks
`expression::areIndependent(modifiedPaths, pathsOfInterest)` when
traversing in either direction and only checks for `renames` if the
`paths` _are_ independent, expecting "independence" to be the same in
either direction.

As rename via `"$addFields"` overwrites a field, it is not actually
symmetric. For the above example:

```
A1 ─────┬──► A2
        │      
        │      
B1 ──►X └──► B2
```

Tracing `"b"` backwards and forwards _should_ differ in whether this
`$addFields` is considered independent or not.

That is, the following are equivalent

```
[{"$addFields":{"b":"$a"}}, {"$match":{"a":123}}]
[{"$match":{"a":123}}, {"$addFields":{"b":"$a"}}]
```

and so are:
```
[{"$addFields":{"b":"$a"}}, {"$match":{"b":123}}]
[{"$match":{"a":123}}, {"$addFields":{"b":"$a"}}]
```

(Including `"b"` in `paths` would currently prevent optimisation by
pushing `$match` earlier in the pipeline)

In contrast, the following can't be similarly expressed with a different
ordering; the value of `"b"` will not be preserved after `$addFields`.

```
[{"$match":{"b":123}}, {"$addFields":{"b":"$a"}}]
```


The minimal change to resolve `$addFields` behaviour is to make
`semantic_analysis::renamedPaths` aware of the destination of renames;
`"b"` is considered modified on a forward pass as it is overwritten, but
is not on a backward pass; the same value is still available elsewhere
before the stage.