SERVER-87951 Modify the Mongo() JS test constructor to optionally create a new connection on each construct (#26628)
This ticket is based on work from
https://github.com/10gen/mongo/pull/26623.

Adds new configuration object to `new Mongo()` that accepts TLS
configuration parameters on the third parameter.
This only adds an option `createNewConnection` so it will look something
like this:

> new Mongo(host, undefined, undefined, {
>         tls: {
>             createNewConnection: true
>         }
>     });

This will set `transientSSLParams->tlsOverrideGlobalParams` to true and
cause the SSLManager to override the global params and creating a new
independent connection.

[SERVER-87952](https://jira.mongodb.org/browse/SERVER-87952) will add
the ability to configure TLS options in the `new Mongo()`