SERVER-88705 Implement histogram CE main interface estimateCardinality (#26762)
SERVER-88705

Implement histogram CE top-level interface. The following methods are
implemented:
1. estimateCardinality
2. canEstimateInterval

The implementation assumes that the interval is of the same type and
requires the types are estimable by either histograms. For example, the
following intervals are supported:
1. [1, 10] // Can estimate via histogram
2. [1, Inf.0] // Can estimate via histogram

The following intervals are not supported in this PR and will be
supported in the follow-up tickets
1. [NaN, NaN] // Type counts
[SERVER-91639](https://jira.mongodb.org/browse/SERVER-91639)
2. [123, "") // Type-backeted interval
[SERVER-94856](https://jira.mongodb.org/browse/SERVER-94856)
3. [null, true] // Non-type-backeted mixed-type interval
[SERVER-94855](https://jira.mongodb.org/browse/SERVER-94855)
1. [false, Date::min()) // Non-type-backeted mixed-type interval
[SERVER-94855](https://jira.mongodb.org/browse/SERVER-94855)
5. ["abc", {a: 1}] // Will not be supported.