SERVER-92586 Ensure $sample size is positive (#26056)
Implements SERVER-92586

This PR changes the behavior of the `$sample` aggregation operation when
invoking it with a `$size` value of `0`.

The documentation for the $sample command currently states:

    The `$sample` stage has the following syntax:
    { $sample: { size: <positive integer N> } }

However, `$sample` commands could previously be invoked with a `$size`
value of `0`, which is not contained in the set of positive integers.

Running a `$sample` command with a `$size` value of `0` returned an
empty result when not running explain on the command.

When using `explain`, invoking a `$sample` command with `$size` `0`
produced a server error, as stated in
https://jira.mongodb.org/browse/SERVER-92320:

foo.aggregate([ {'$count': 'num'}, {'$sample': {'size':
0}}]).explain("executionStats")
MongoServerError[Location5466400]: Invalid call to memory usage tracker,
could not find function num

So invoking the `$sample` command with a size of `0` was working in some
cases, but not in all.

The PR changes the behavior so that invoking the `$sample` command with
a `$size` value of `0` will now always fail with uassert(28747).
This makes the server behave as documented. Using a `$sample` size of
`0` previously wasn't supported according to the documentation.

The error code 28747 previously already existed and was used when a
negative sample size was requested, and it is reused now to complain
about negative and zero sample sizes.
This PR slightly changes the returned error message for error code 28747
from `size argument to $sample must not be negative` to `size argument
to $sample must be a positive integer`. This may affect drivers or
client applications that rely on the exact error message for error code
28747.