SERVER-74543 Remove 'backgroundSecondary' and 'head' fields in index catalog metadata (#27216)
The `head` field only exists to support downgrade compatibility to 4.0,
which is no longer possible.

The `backgroundSecondary` field only exists for downgrade compatibility
to 4.2 where two-phase index builds (where a `buildUUID` is always
present) are not always used. The only case where `backgroundSecondary`
could be acted on today is if an index build is unfinished during
initial sync, where `buildUUID` is not present, and the server restarts.
Before the change, the index build would be rebuilt, but it should
actually be dropped, as it will be dropped anyway when the initial sync
restarts.

If `backgroundSecondary` is not persisted during initial sync, the
server crashes, and is down downgraded to 8.0, a missing
`backgroundSecondary` will be interpreted as 'true', and the same,
existing logic is activated.