SERVER-94879 Clean up concurrency suite exclusion lists (#27231)
In this PR, I:
* Removed `reindex.*.js` from each suite and tagged it with
`requires_standalone`. Additionally, I added `requires_standalone` to
every suite.
* Removed `agg_group_external.js` and `agg_sort_external.js` from every
suite and tagged them with `requires_standalone`.
* Removed `findAndModify_update_grow.js` from being explicitly excluded
and give it the `requires_standalone` tag.
* Removed any test that is "double excluded", i.e. tests excluded from a
suite via exclude files and via exclude tags

I ran a script that I wrote to make sure that the set of tests being run
in the old world and post cleanup are the same.