SERVER-94206 Log actual validation options not user-provided command flags (#27040)
# Issue

SERVER-94206

# Description

The user-provided commands to validate don't quite align with the
options as-used by validation code, e.g. some options not provided by
the user will be inferred from others.

In this change, we make the log line match the actual parameters given
to validate, rather than the ones the user provided.

To facilitate this change, we take a stricter stance on the difference
between "preconditions" of validation (like the options the user
provides to it) and "runtime state" of validation. In practice, that
means separating all of the immutable-during-validation values into
their own class.

---------

Co-authored-by: Gregory Noma <gregory.noma@gmail.com>