SERVER-89708 re-enable timeseries_mixed_operations and retry in case of NoProgressMade #27063 (#27064)
The test attempts to insert several documents with both ordered and
ordered insertion into a timeseries and verifies the insertion worked.

With moveCollection in the background, the insertion might stop behind
the reshard critical section and fail with NoProgressMade if waiting for
too long.

When running on a slow varian, this case is likely.