SERVER-71453 Change count command from BasicCommand to IDL typed command (#26579)
Change the type of the previously existing `CmdCount` class from
`BasicCommand` to IDL-based `CountCmdVersion1Gen<CmdCount>`.

The reply type for the count command's `n` field is declared as
`variant: [int, long]` to keep backwards-compatibility with client
applications and drivers, which may not all handle a long value
correctly.

As `count` is now an IDL-typed command, an additional error code can now
be returned if the NSS/UUID in the request does not have the expected
type. When not passing in a valid NSS or UUID as input for the count
command, the command previously return the `ErrorCodes.InvalidNamespace`
error. Now, with the IDL validation, the `count` command will now return
the `ErrorCodes::TypeMismatch` if the request field is not of type NSS
or UUID. This is a downwards-incompatible change in error codes. The
error code does not change for well-formed NSS or UUID values which
point to non-existing databases/collections.