SERVER-94861 change_streams_shards_start_in_sync.js should temporarily hang shard1 instead of rejecting cursor establishing request from mongos (#27170)
change_streams_shards_start_in_sync.js temporarily disconnected shard1
from the mongos so that it could test that change streams on sharded
collections start in sync with each other. The issue with this was that
this would cause the $changeStream request on the mongos to fail because
when the mongos attempted to establish a cursor on shard1, it would get
a HostUnreachable error. Instead of rejecting the cursor establishing
request from the mongos, shard1 should temporarily hang.