SERVER-93113 Implement ExhaustResponseReader (#27323)
This commit introduces the `NetworkInterface::ExhaustResponseReader`
type, which will be returned from exhaust-command-starting methods on
`NetworkInterface`. This commit also introduces an implementation,
`ExhaustResponseReaderTL`, which will be used by `NetworkInterfaceTL`.

This commit also includes the changes from SERVER-94408, which fixed a
bug in the added tests when this was originally merged to a feature
branch.

---------

Co-authored-by: Joseph Prince <57415450+josephdprince@users.noreply.github.com>