SERVER-95102 Raise timeout of bazel_coverage task (#27358)
After converting more targets to Bazel this task is now hitting the 40
minute execution limit in https://jira.mongodb.org/browse/BF-35144.
Raise the limit to 180 minutes. This will likely need to be optimized to
either use remote execution or skip sandbox recreation later, but for
now let's increase the limit to unblock the Bazel target migration and
keep the task running successfully.