SERVER-93183: Remove queryPlannerDebug verbosity (#27171)
This PR contains:
1. A pretty straightforward revert of
[SERVER-83121](https://jira.mongodb.org/browse/SERVER-83121) to remove
the queryPlannerDebug verbosity
2. An update to the IDL compatibility checker to ensure that this change
is considered safe (i.e., does not break compatibility). The rationale
for this change is that we never released this new verbosity to users;
it was hidden behind a disabled feature flag. No one should have been
using this verbosity, so removing it should be safe.