SERVER-94465: Update Kafka consumer group offset in KafkaConsumerOperator::doRunOnce (#27004)
## Description

This PR contains changes add a configuration option that changes the
behavior of KafkaConsumerOperator.

When enabled, the operator will update the Kafka consumer group's offset
as it receives input data.

When disabled (default), it will display the current behavior where it
is only updated whenever a checkpoint occurs which can range from every
5 minutes to an hour (with windows).

https://jira.mongodb.org/browse/SERVER-94465