SERVER-95112 Add `PlanStage*` member to `PlanStageStats` (#27384)
This necessary for the implementation of explain in the cost-based
ranker. CBR maintains the following mappings:
* `QuerySolutionNode*` to Estimate
* `PlanStage*` to `QuerySolutionNode*`

When we run explain, we invoke `getStats()` on the root of the
`PlanStage` tree to get a `PlanStageStats` tree with the same topology.
We then use this tree to construct a BSON output of the explain. In
order to annoate explain nodes with their corresponding estimates, we
need to get the `PlanStage*` associated with a particular
`PlanStageStats` node and consult the two mappings described above.