SERVER-94826 Make multi-key-index repair off by default and add it as an option (#27408)
When repair is false, we still fix multi keys as this is on by default.
We should have this off by default and only applied when either repair
is true or it is called explicitly, an option for fixMultikey was
created to address this. Tests were modified to reflect these changes,
and a relevant log message was adjusted.