SERVER-85864 Use 64 bit signed integer for num_hosts_targeted_metrics (#27491)
This changes 32 bit integers for num_hosts_targeted_metrics to 64 bit
integers.

The reason why 64 unsigned integers have not been used like the ticket
mention is that 64 bit unsigned numbers are only used for timestamps and
allowing all 64 bit unsigned numbers would require changes in
`BSONObjBuilderBase`. We decided to avoid those changes.

Signed 64 bit integers were deemed good enough to fix the issue that was
described by the jira ticket SERVER-85864

In the code `long long` was used since on platform that both `long` and
`long long` are 64 bit, they are still different fundamental types.
Since `std::int64_t` could be defined as `long int`, it would not find
the right template specialization for `BSONObjAppendFormat` and the
right overload for `appendNumber`. The only good way is to use the same
type as `BSONObjBuilderBase` uses.